move relevant models into hecTaxCheck sub-package
