+++ app/uk/gov/hmrc/hec/models/TaxCheckListItem.scala
@@ -18,6 +18,7 @@ package uk.gov.hmrc.hec.models
 
 import play.api.libs.json.{Json, OFormat}
 import uk.gov.hmrc.hec.models.licence.LicenceType
+import uk.gov.hmrc.hec.models.hecTaxCheck.{HECTaxCheck, HECTaxCheckCode}
 
 import java.time.{LocalDate, ZonedDateTime}
