+++ app/uk/gov/hmrc/hec/models/hecTaxCheck/LicenceDetails.scala
@@ -14,14 +14,19 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.hec.models
+package uk.gov.hmrc.hec.models.hecTaxCheck
 
-sealed trait CTLookupStatus extends Product with Serializable
+import play.api.libs.json.{Json, OFormat}
+import uk.gov.hmrc.hec.models.licence.{LicenceTimeTrading, LicenceType, LicenceValidityPeriod}
 
-object CTLookupStatus {
+final case class LicenceDetails(
+  licenceType: LicenceType,
+  licenceTimeTrading: LicenceTimeTrading,
+  licenceValidityPeriod: LicenceValidityPeriod
+)
 
-  case object Successful extends CTLookupStatus
+object LicenceDetails {
 
-  case object NoLiveRecords extends CTLookupStatus
+  implicit val format: OFormat[LicenceDetails] = Json.format
 
 }
