+++ app/uk/gov/hmrc/hec/repos/HECTaxCheckStore.scala
@@ -25,7 +25,8 @@ import org.mongodb.scala.model.{Filters, IndexModel, IndexOptions, Indexes, Sort
 import play.api.Configuration
 import play.api.libs.json.{JsError, JsSuccess, JsonValidationError}
 import uk.gov.hmrc.hec.models.ids.GGCredId
-import uk.gov.hmrc.hec.models.{Error, HECTaxCheck, HECTaxCheckCode}
+import uk.gov.hmrc.hec.models.hecTaxCheck.{HECTaxCheck, HECTaxCheckCode}
+import uk.gov.hmrc.hec.models.Error
 import uk.gov.hmrc.hec.util.Logging
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.mongo.cache.{CacheIdType, CacheItem, DataKey, MongoCacheRepository}
