+++ app/uk/gov/hmrc/hec/services/FileCreationService.scala
@@ -20,9 +20,6 @@ import cats.implicits.catsSyntaxOptionId
 import cats.syntax.eq._
 import com.google.inject.{ImplementedBy, Inject}
 import uk.gov.hmrc.hec.models.CorrectiveAction._
-import uk.gov.hmrc.hec.models.HECTaxCheckData.{CompanyHECTaxCheckData, IndividualHECTaxCheckData}
-import uk.gov.hmrc.hec.models.HECTaxCheckSource.Digital
-import uk.gov.hmrc.hec.models.SAStatus._
 import uk.gov.hmrc.hec.models.TaxSituation._
 import uk.gov.hmrc.hec.models.fileFormat.FileFormat.toFileContent
 import uk.gov.hmrc.hec.models.fileFormat._
@@ -30,7 +27,11 @@ import uk.gov.hmrc.hec.models.licence.LicenceTimeTrading.{EightYearsOrMore, Four
 import uk.gov.hmrc.hec.models.licence.LicenceType.{DriverOfTaxisAndPrivateHires, OperatorOfPrivateHireVehicles, ScrapMetalDealerSite, ScrapMetalMobileCollector}
 import uk.gov.hmrc.hec.models.licence.LicenceValidityPeriod.{UpToFiveYears, UpToFourYears, UpToOneYear, UpToThreeYears, UpToTwoYears}
 import uk.gov.hmrc.hec.models.licence.{LicenceTimeTrading, LicenceType, LicenceValidityPeriod}
-import uk.gov.hmrc.hec.models.{CTStatus, CorrectiveAction, Error, HECTaxCheck, HECTaxCheckFileBodyList, SAStatus, TaxSituation, YesNoAnswer}
+import uk.gov.hmrc.hec.models.hecTaxCheck.HECTaxCheckData.{CompanyHECTaxCheckData, IndividualHECTaxCheckData}
+import uk.gov.hmrc.hec.models.hecTaxCheck.HECTaxCheckSource.Digital
+import uk.gov.hmrc.hec.models.hecTaxCheck.SAStatus._
+import uk.gov.hmrc.hec.models.hecTaxCheck.{CTStatus, HECTaxCheck, HECTaxCheckFileBodyList, SAStatus}
+import uk.gov.hmrc.hec.models.{CorrectiveAction, Error, TaxSituation, YesNoAnswer}
 import uk.gov.hmrc.hec.util.TimeProvider
 
 import java.time.ZoneId
