+++ app/uk/gov/hmrc/hec/testonly/controllers/TaxCheckController.scala
@@ -22,9 +22,9 @@ import cats.syntax.eq._
 import com.google.inject.{Inject, Singleton}
 import play.api.libs.json.{JsError, JsSuccess, JsValue, Json}
 import play.api.mvc.{Action, AnyContent, ControllerComponents}
-import uk.gov.hmrc.hec.models.HECTaxCheckCode
 import uk.gov.hmrc.hec.testonly.services.TaxCheckService
 import uk.gov.hmrc.hec.controllers.toFuture
+import uk.gov.hmrc.hec.models.hecTaxCheck.HECTaxCheckCode
 import uk.gov.hmrc.hec.testonly.models.SaveTaxCheckRequest
 import uk.gov.hmrc.hec.util.Logging
 import uk.gov.hmrc.hec.util.Logging._
