+++ test/uk/gov/hmrc/hec/controllers/IFControllerSpec.scala
@@ -23,8 +23,9 @@ import play.api.inject.guice.GuiceableModule
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import uk.gov.hmrc.auth.core.AuthConnector
-import uk.gov.hmrc.hec.models._
+import uk.gov.hmrc.hec.models.{hecTaxCheck, _}
 import uk.gov.hmrc.hec.models.ids.{CTUTR, GGCredId, SAUTR}
+import uk.gov.hmrc.hec.models.hecTaxCheck.{CTStatusResponse, SAStatus, SAStatusResponse}
 import uk.gov.hmrc.hec.services.IFService
 import uk.gov.hmrc.hec.services.IFService.{BackendError, DataNotFoundError, IFError}
 import uk.gov.hmrc.http.HeaderCarrier
@@ -131,7 +132,7 @@ class IFControllerSpec extends ControllerSpec with AuthSupport {
           val utr        = SAUTR(validSautr)
           val taxYearStr = "2020"
           val taxYear    = TaxYear(taxYearStr.toInt)
-          val response   = SAStatusResponse(
+          val response   = hecTaxCheck.SAStatusResponse(
             sautr = utr,
             taxYear = taxYear,
             status = SAStatus.ReturnFound
