+++ test/uk/gov/hmrc/hec/services/IFServiceImplSpec.scala
@@ -29,7 +29,8 @@ import play.api.libs.json.Json
 import play.api.test.Helpers._
 import uk.gov.hmrc.hec.connectors.IFConnector
 import uk.gov.hmrc.hec.models.ids.{CTUTR, SAUTR}
-import uk.gov.hmrc.hec.models.{CTAccountingPeriod, CTStatus, CTStatusResponse, Error, SAStatus, SAStatusResponse, TaxYear}
+import uk.gov.hmrc.hec.models.hecTaxCheck.{CTAccountingPeriod, CTStatus, CTStatusResponse, SAStatus, SAStatusResponse}
+import uk.gov.hmrc.hec.models.{Error, TaxYear}
 import uk.gov.hmrc.hec.services.IFService.{BackendError, DataNotFoundError}
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
