+++ test/uk/gov/hmrc/hec/services/TaxCheckServiceImplSpec.scala
@@ -24,12 +24,14 @@ import org.scalamock.scalatest.MockFactory
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import play.api.test.Helpers._
-import uk.gov.hmrc.hec.models.ApplicantDetails.{CompanyApplicantDetails, IndividualApplicantDetails}
-import uk.gov.hmrc.hec.models.HECTaxCheckData.{CompanyHECTaxCheckData, IndividualHECTaxCheckData}
-import uk.gov.hmrc.hec.models.TaxDetails.{CompanyTaxDetails, IndividualTaxDetails}
+import uk.gov.hmrc.hec.models
+import uk.gov.hmrc.hec.models.hecTaxCheck.ApplicantDetails.{CompanyApplicantDetails, IndividualApplicantDetails}
+import uk.gov.hmrc.hec.models.hecTaxCheck.HECTaxCheckData.{CompanyHECTaxCheckData, IndividualHECTaxCheckData}
+import uk.gov.hmrc.hec.models.hecTaxCheck.TaxDetails.{CompanyTaxDetails, IndividualTaxDetails}
 import uk.gov.hmrc.hec.models.ids._
-import uk.gov.hmrc.hec.models.licence.{LicenceDetails, LicenceTimeTrading, LicenceType, LicenceValidityPeriod}
-import uk.gov.hmrc.hec.models.{CTAccountingPeriod, CTStatus, CTStatusResponse, CompanyHouseName, DateOfBirth, Error, HECTaxCheck, HECTaxCheckCode, HECTaxCheckData, HECTaxCheckMatchRequest, HECTaxCheckMatchResult, HECTaxCheckMatchStatus, HECTaxCheckSource, Name, TaxCheckListItem, TaxSituation, TaxYear, YesNoAnswer}
+import uk.gov.hmrc.hec.models.licence.{LicenceTimeTrading, LicenceType, LicenceValidityPeriod}
+import uk.gov.hmrc.hec.models.hecTaxCheck.{CTAccountingPeriod, CTStatus, CTStatusResponse, HECTaxCheck, HECTaxCheckCode, HECTaxCheckData, HECTaxCheckMatchRequest, HECTaxCheckMatchResult, HECTaxCheckMatchStatus, HECTaxCheckSource, LicenceDetails}
+import uk.gov.hmrc.hec.models.{CompanyHouseName, DateOfBirth, Error, Name, TaxCheckListItem, TaxSituation, TaxYear, YesNoAnswer, hecTaxCheck}
 import uk.gov.hmrc.hec.repos.HECTaxCheckStore
 import uk.gov.hmrc.hec.util.{TimeProvider, TimeUtils}
 import uk.gov.hmrc.http.HeaderCarrier
@@ -129,7 +131,7 @@ class TaxCheckServiceImplSpec extends AnyWordSpec with Matchers with MockFactory
       val expectedExpiryDate                        = TimeUtils.today().plusDays(expiresAfter.toDays)
       val taxCheckCode                              = HECTaxCheckCode("code")
       def taxCheck(fileCorrelationId: Option[UUID]) =
-        HECTaxCheck(taxCheckData, taxCheckCode, expectedExpiryDate, now, false, fileCorrelationId)
+        models.hecTaxCheck.HECTaxCheck(taxCheckData, taxCheckCode, expectedExpiryDate, now, false, fileCorrelationId)
 
       "return an error" when {
 
@@ -191,14 +193,14 @@ class TaxCheckServiceImplSpec extends AnyWordSpec with Matchers with MockFactory
         CRN("crn") -> CRN("incorrect-crn")
 
       val storedLicenceDetails =
-        LicenceDetails(
+        hecTaxCheck.LicenceDetails(
           storedLicenceType,
           LicenceTimeTrading.EightYearsOrMore,
           LicenceValidityPeriod.UpToOneYear
         )
 
       val storedIndividualTaxCheck =
-        HECTaxCheck(
+        hecTaxCheck.HECTaxCheck(
           IndividualHECTaxCheckData(
             IndividualApplicantDetails(Some(GGCredId("")), Name("", ""), storedDateOfBirth),
             storedLicenceDetails,
@@ -222,7 +224,7 @@ class TaxCheckServiceImplSpec extends AnyWordSpec with Matchers with MockFactory
         )
 
       val storedCompanyTaxCheck =
-        HECTaxCheck(
+        hecTaxCheck.HECTaxCheck(
           CompanyHECTaxCheckData(
             CompanyApplicantDetails(GGCredId("").some, storedCRN, CompanyHouseName("Test Tech Ltd")),
             storedLicenceDetails,
@@ -431,7 +433,7 @@ class TaxCheckServiceImplSpec extends AnyWordSpec with Matchers with MockFactory
 
       val taxCheckData = CompanyHECTaxCheckData(
         CompanyApplicantDetails(ggCredId.some, CRN(""), CompanyHouseName("Test Tech Ltd")),
-        LicenceDetails(
+        hecTaxCheck.LicenceDetails(
           LicenceType.ScrapMetalDealerSite,
           LicenceTimeTrading.EightYearsOrMore,
           LicenceValidityPeriod.UpToOneYear
@@ -473,9 +475,9 @@ class TaxCheckServiceImplSpec extends AnyWordSpec with Matchers with MockFactory
         val code2 = HECTaxCheckCode("code2")
         val code3 = HECTaxCheckCode("code3")
 
-        val taxCheckToday     = HECTaxCheck(taxCheckData, code1, today, now, false, None)
-        val taxCheckYesterday = HECTaxCheck(taxCheckData, code2, yesterday, now, false, None)
-        val taxCheckTomorrow  = HECTaxCheck(taxCheckData, code3, tomorrow, now, false, None)
+        val taxCheckToday     = hecTaxCheck.HECTaxCheck(taxCheckData, code1, today, now, false, None)
+        val taxCheckYesterday = hecTaxCheck.HECTaxCheck(taxCheckData, code2, yesterday, now, false, None)
+        val taxCheckTomorrow  = hecTaxCheck.HECTaxCheck(taxCheckData, code3, tomorrow, now, false, None)
 
         val todayItem    = TaxCheckListItem(
           taxCheckToday.taxCheckData.licenceDetails.licenceType,
@@ -503,7 +505,7 @@ class TaxCheckServiceImplSpec extends AnyWordSpec with Matchers with MockFactory
 
       val taxCheckData: HECTaxCheckData = CompanyHECTaxCheckData(
         CompanyApplicantDetails(ggCredId.some, CRN(""), CompanyHouseName("Test Tech Ltd")),
-        LicenceDetails(
+        hecTaxCheck.LicenceDetails(
           LicenceType.ScrapMetalDealerSite,
           LicenceTimeTrading.EightYearsOrMore,
           LicenceValidityPeriod.UpToOneYear
@@ -540,11 +542,11 @@ class TaxCheckServiceImplSpec extends AnyWordSpec with Matchers with MockFactory
       "only return HEC Tac check code with isExtracted false" in {
 
         val hecTaxCheck1 =
-          HECTaxCheck(taxCheckData, HECTaxCheckCode("ABC 123 ABC"), today.plusDays(1), now, false, None)
+          hecTaxCheck.HECTaxCheck(taxCheckData, HECTaxCheckCode("ABC 123 ABC"), today.plusDays(1), now, false, None)
         val hecTaxCheck2 =
-          HECTaxCheck(taxCheckData, HECTaxCheckCode("EBC 123 ABC"), today.plusDays(1), now, false, None)
+          hecTaxCheck.HECTaxCheck(taxCheckData, HECTaxCheckCode("EBC 123 ABC"), today.plusDays(1), now, false, None)
         val hecTaxCheck3 =
-          HECTaxCheck(taxCheckData, HECTaxCheckCode("MBC 123 ABC"), today.plusDays(1), now, false, None)
+          hecTaxCheck.HECTaxCheck(taxCheckData, HECTaxCheckCode("MBC 123 ABC"), today.plusDays(1), now, false, None)
 
         mockGetAllTaxCheckCodesByStatus(false, 0, 3, "_id")(Right(List(hecTaxCheck1, hecTaxCheck2, hecTaxCheck3)))
 
@@ -559,7 +561,7 @@ class TaxCheckServiceImplSpec extends AnyWordSpec with Matchers with MockFactory
 
       val taxCheckData: HECTaxCheckData = CompanyHECTaxCheckData(
         CompanyApplicantDetails(ggCredId.some, CRN(""), CompanyHouseName("Test Tech Ltd")),
-        LicenceDetails(
+        hecTaxCheck.LicenceDetails(
           LicenceType.ScrapMetalDealerSite,
           LicenceTimeTrading.EightYearsOrMore,
           LicenceValidityPeriod.UpToOneYear
@@ -596,9 +598,23 @@ class TaxCheckServiceImplSpec extends AnyWordSpec with Matchers with MockFactory
       "only return HEC Tac check code with the given CorrelationId" in {
 
         val hecTaxCheck1 =
-          HECTaxCheck(taxCheckData, HECTaxCheckCode("ABC 123 ABC"), today.plusDays(1), now, false, uuid.some)
+          hecTaxCheck.HECTaxCheck(
+            taxCheckData,
+            HECTaxCheckCode("ABC 123 ABC"),
+            today.plusDays(1),
+            now,
+            false,
+            uuid.some
+          )
         val hecTaxCheck2 =
-          HECTaxCheck(taxCheckData, HECTaxCheckCode("EBC 123 ABC"), today.plusDays(1), now, false, uuid.some)
+          hecTaxCheck.HECTaxCheck(
+            taxCheckData,
+            HECTaxCheckCode("EBC 123 ABC"),
+            today.plusDays(1),
+            now,
+            false,
+            uuid.some
+          )
 
         mockGetAllTaxCheckCodesByCorrelationId(uuid)(Right(List(hecTaxCheck1, hecTaxCheck2)))
